Day 4 SavePDFFull page is now ready.
made the javascript and modified manifest ot gain permission to access
current tab and then moved the html with the custom styling to a popup
window where the user should print it normally.

modified styleHTML.css to improve readability of pdf lines now
break-line is set to anywhere instead of auto.

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
      modified:   Trial/index.html
      modified:   css/styleHTML.css
      deleted:    js/index.js
      new file:   js/main.js
      modified:   manifest.json
      modified:   popup.html

Untracked files:
      Trial/allHTMLElements.html
      Trial/real-currentChat.html
      manifest-removeed.json
      media/images/jpg.jpg